requirements.txt
ultralytics
gradio
opencv-python
Pillow
